LayoutTests/editing/execCommand/queryCommandValue-unsupported-commands.html

<!DOCTYPE html>
<html>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<div id="test" contenteditable></div>
<script>
description("Tests queryCommandValue returns empty string for unsupported commands");

var test = document.getElementById('test');
test.focus();
document.execCommand('fontSize', false, 5);
shouldBe("document.queryCommandValue('foofoo')", "''");
shouldBe("typeof document.queryCommandValue('foofoo')", "'string'");
shouldBe("document.queryCommandValue('fontSize')", "'5'");
shouldBe("typeof document.queryCommandValue('fontSize')", "'string'");
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</html>
