third_party/WebKit/LayoutTests/editing/execCommand/queryCommandValue-unsupported-commands-expected.txt

Tests queryCommandValue returns empty string for unsupported commands

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.queryCommandValue('foofoo') is ''
PASS typeof document.queryCommandValue('foofoo') is 'string'
PASS document.queryCommandValue('fontSize') is '5'
PASS typeof document.queryCommandValue('fontSize') is 'string'
PASS successfullyParsed is true

TEST COMPLETE
